+++ bill_extractor/providers/__init__.py
@@ -1,9 +1,10 @@
 """Registered electricity-provider adapters."""
 
+from .dgvcl import DGVCLParser
 from .pgvcl import PGVCLParser
 from .torrent import TorrentParser
 from .ugvcl import UGVCLParser
 
-PARSERS = (UGVCLParser(), TorrentParser(), PGVCLParser())
+PARSERS = (UGVCLParser(), TorrentParser(), PGVCLParser(), DGVCLParser())
 
 __all__ = ["PARSERS"]